+++ crates/js-sys/src/lib.rs
@@ -27,10 +27,7 @@
 #![doc(html_root_url = "https://docs.rs/js-sys/0.2")]
 #![cfg_attr(not(feature = "std"), no_std)]
 #![cfg_attr(target_feature = "atomics", feature(thread_local))]
-#![cfg_attr(
-    all(feature = "futures", target_feature = "atomics"),
-    feature(stdarch_wasm_atomic_wait)
-)]
+#![cfg_attr(target_feature = "atomics", feature(stdarch_wasm_atomic_wait))]
 
 extern crate alloc;
 
@@ -13896,7 +13893,6 @@ arrays! {
 
 /// Bridging between JavaScript `Promise`s and Rust `Future`s.
 ///
-/// Enables `promise.await` directly on any [`Promise`] when this feature is active.
-/// This module is automatically available when depending on `wasm-bindgen-futures`.
-#[cfg(feature = "futures")]
+/// Enables `promise.await` directly on any [`Promise`].
+/// This module is also re-exported by `wasm-bindgen-futures` for backwards compatibility.
 pub mod futures;