+++ crates/macro-support/src/parser.rs
@@ -798,6 +798,7 @@ impl<'a> ConvertToAst<(&ast::Program, BindgenAttrs, &'a Option<ast::ImportModule
             doc_comment,
             wasm_bindgen: program.wasm_bindgen.clone(),
             wasm_bindgen_futures: program.wasm_bindgen_futures.clone(),
+            js_sys: program.js_sys.clone(),
             generics: self.sig.generics,
         });
         opts.check_used();
@@ -1440,6 +1441,7 @@ impl<'a> MacroParse<(Option<BindgenAttrs>, &'a mut TokenStream)> for syn::Item {
                     start,
                     wasm_bindgen: program.wasm_bindgen.clone(),
                     wasm_bindgen_futures: program.wasm_bindgen_futures.clone(),
+                    js_sys: program.js_sys.clone(),
                 });
             }
             syn::Item::Impl(mut i) => {
@@ -1572,13 +1574,14 @@ fn prepare_for_impl_recursion(
 
     let wasm_bindgen = &program.wasm_bindgen;
     let wasm_bindgen_futures = &program.wasm_bindgen_futures;
+    let js_sys = &program.js_sys;
     method.attrs.insert(
         0,
         syn::Attribute {
             pound_token: Default::default(),
             style: syn::AttrStyle::Outer,
             bracket_token: Default::default(),
-            meta: syn::parse_quote! { #wasm_bindgen::prelude::__wasm_bindgen_class_marker(#class = #js_class, wasm_bindgen = #wasm_bindgen, wasm_bindgen_futures = #wasm_bindgen_futures) },
+            meta: syn::parse_quote! { #wasm_bindgen::prelude::__wasm_bindgen_class_marker(#class = #js_class, wasm_bindgen = #wasm_bindgen, wasm_bindgen_futures = #wasm_bindgen_futures, js_sys = #js_sys) },
         },
     );
 
@@ -1594,10 +1597,12 @@ impl MacroParse<&ClassMarker> for &mut syn::ImplItemFn {
             js_class,
             wasm_bindgen,
             wasm_bindgen_futures,
+            js_sys,
         }: &ClassMarker,
     ) -> Result<(), Diagnostic> {
         program.wasm_bindgen = wasm_bindgen.clone();
         program.wasm_bindgen_futures = wasm_bindgen_futures.clone();
+        program.js_sys = js_sys.clone();
 
         match self.vis {
             syn::Visibility::Public(_) => {}
@@ -1663,6 +1668,7 @@ impl MacroParse<&ClassMarker> for &mut syn::ImplItemFn {
             start: ast::StartKind::None,
             wasm_bindgen: program.wasm_bindgen.clone(),
             wasm_bindgen_futures: program.wasm_bindgen_futures.clone(),
+            js_sys: program.js_sys.clone(),
         });
         opts.check_used();
         Ok(())
@@ -2442,12 +2448,18 @@ fn main(program: &ast::Program, mut f: ItemFn, tokens: &mut TokenStream) -> Resu
 
     let wasm_bindgen = &program.wasm_bindgen;
     let wasm_bindgen_futures = &program.wasm_bindgen_futures;
+    let js_sys = &program.js_sys;
+    let futures = if ast::use_js_sys_futures() {
+        quote::quote! { #js_sys::futures }
+    } else {
+        quote::quote! { #wasm_bindgen_futures }
+    };
 
     if f.sig.asyncness.take().is_some() {
         *f.block = syn::parse2(quote::quote! {
                 {
                     async fn __wasm_bindgen_generated_main() #r#return #body
-                    #wasm_bindgen_futures::spawn_local(
+                    #futures::spawn_local(
                         async move {
                             use #wasm_bindgen::__rt::Main;
                             let __ret = __wasm_bindgen_generated_main();